Refactoring: Extract reference file builder to a mixin
