+++ lib/rmt/rpm/deltainfo_xml_parser.rb
@@ -1,15 +1,11 @@
 class RMT::Rpm::DeltainfoXmlParser < RepomdParser::DeltainfoXmlParser
-  attr_reader :referenced_files
-  def initialize(filename, mirror_src = false)
-    super(filename)
-    @mirror_src = mirror_src
-  end
+  include UtilsMixin
 
   def end_element(name)
     return unless delta? name
 
     unless delta_src_package_with_mirror_src?
-      build_reference_for_package_with_src(@delta)
+      build_reference_for_package_with_src(@delta, :drpm)
       return
     end
 
@@ -24,15 +20,5 @@ class RMT::Rpm::DeltainfoXmlParser < RepomdParser::DeltainfoXmlParser
 
   def delta_src_package_with_mirror_src?
     (@package[:arch] == 'src' && !@mirror_src)
-  end
-
-  def build_reference_for_package_with_src(delta)
-    @referenced_files << RepomdParser::Reference.new(
-      location: delta[:location],
-       checksum_type: delta[:checksum_type],
-       checksum: delta[:checksum],
-       type: :drpm,
-       size: delta[:size].to_i || nil
-     )
   end
 end
